Silence most compiler warnings
These include:
C4018 - Comparing signed and unsigned integer. This most commonly occurs
        when we compare a signed integer against some_stl_component.size()
C4290 - exception specifications. Visual Studio 2010 does not support
        this, so we silence its warnings
C4101 - Unreferenced local variable
C4996 - Using "deprecated" functions such as strcpy, strcat, fopen

The only other major class of warnings that have not been silenced is
C4715, where a function/method does not return a value at all paths.
This will be handled in a future patch.

+++ SPA/ResultsProjector.cpp
@@ -197,7 +197,7 @@ void ResultsProjector::get_results(ResultsTable& resultsTable,
 void ResultsProjector::recursive_generate(int n,
         QueryInfo *qinfo, PKB *pkb)
 {
-    if (n >= this->tablesUsed.size()) {
+    if (n >= (int)this->tablesUsed.size()) {
         // generate results
         const vector<AttrRef>& selectTuple = qinfo->get_selectTuple();
         this->sb.clear();
